Update rails to 5.2 (incl. bootsnap)
This comment might contain some changes that should have been included
in the one containing the update to rails 5.1, but unfortunately I can't
debug which those are.

The encrypted secrets feature is not used by schleuder-web, neither is
the built-in support for custom configuration.

I tried to migrate the configuration so that the behaviour of the
application doesn't change through the upgrades to 5.x.