production: set log level to warn, instead of debug
This way, IP addresses won't be logged, which is reasonable for a
software like schleuder-web. Also, this saves quite some space due to
the reduced logging. Still, if required, people are able to increase the
logging manually if there is a need to, or they prefer more verbose
logging in general.

Closes #115