HEADSUP/BREAKING: SECRET_KEY_BASE now required in prod env
It is now mandatory to the the SECRET_KEY_BASE environment variable for
'production'.
Configuring that value in `config/secrets.yml` is not supported anymore
(this is a change coming with Rails v7.1, not my choice).

For the Rails environments 'development' and 'test' a random value is
used by default.

Background information: we're not using Rails' new default way to store
these values (`config/credentials.yml.enc` etc.) because the environment
variable aligns better with known usage of schleuder-web and is easier
to understand.